some cart and business logic

+++ controllers/handlerFactory.js
@@ -1,7 +1,8 @@
 const catchAsync = require("./../utils/catchAsync");
 const AppError = require("./../utils/appError");
 const APIFeatures = require("./../utils/apiFeatures");
-const CartItem = require("../models/cartItemModel");
+const CartItem = require("../models/cartModel");
+const ensureEnoughStock = require("./../utils/ensureEnoughStock");
 
 const sendResponse = function (statusCode, data, res, token, message) {
   const response = {
@@ -17,9 +18,11 @@ const sendResponse = function (statusCode, data, res, token, message) {
 
 exports.sendResponse = sendResponse;
 
-exports.getAll = (Model) =>
+exports.getAll = (Model, options) =>
   catchAsync(async (req, res, next) => {
-    let features = new APIFeatures(Model.find(), req.query).filter().sort().paginate().limitFields();
+    const filter = options?.filterByUser ? { user: req.user.id } : {};
+
+    let features = new APIFeatures(Model.find(filter), req.query).filter().sort().paginate().limitFields();
     const docs = await features.query;
 
     sendResponse(200, docs, res);
@@ -36,24 +39,36 @@ exports.getOne = (Model, populateOptions) =>
 
 exports.createOne = (Model, options) =>
   catchAsync(async (req, res, next) => {
-    let doc;
-    let message;
+    let doc, message;
     let statusCode = 201;
 
-    if (options?.cartItemLogic) {
-      const { user, book } = req.body;
+    if (options?.cartLogic) {
+      const { user, book, quantity = 1 } = req.body;
 
-      doc = await CartItem.findOne({ user, book });
+      doc = await Model.findOne({ user }); // cart
 
-      if (doc) {
-        doc.quantity += req.body.quantity || 1;
-        await doc.save();
-        message = "Quantity updated in cart!";
-        statusCode = 200;
+      if (!doc) {
+        // no cart yet
+        await ensureEnoughStock(book, quantity);
+        doc = await Model.create({ user, items: [{ book, quantity }] });
+        message = "Cart created and item added!";
       } else {
-        doc = await Model.create(req.body);
-        message = "Item added to cart!";
+        const item = doc.items.find((i) => i.book.toString() === book);
+
+        if (item) {
+          const newQuantity = item.quantity + quantity;
+          await ensureEnoughStock(book, newQuantity);
+          item.quantity = newQuantity;
+          message = "Quantity updated in cart!";
+        } else {
+          await ensureEnoughStock(book, quantity);
+          doc.items.push({ book, quantity });
+          message = "Item added to cart!";
+        }
       }
+
+      await doc.save();
+      statusCode = 200;
     } else {
       doc = await Model.create(req.body);
       message = "created successfully!";
@@ -64,6 +79,26 @@ exports.createOne = (Model, options) =>
 
 exports.updateOne = (Model, options) =>
   catchAsync(async (req, res, next) => {
+    //
+    if (options?.cartLogic) {
+      const cart = await Model.findOne({ user: req.user.id });
+      if (!cart) return next(new AppError(404, "Cart not found"));
+
+      const item = cart.items.id(req.params.id);
+      if (!item) return next(new AppError(404, "Item not found"));
+
+      if (!req.body.quantity) return next(new AppError(400, "Please provide a quantity"));
+
+      const book = await mongoose.model("Book").findById(item.book).select("stock");
+      if (!book) return next(new AppError(404, "Book not found"));
+
+      if (book.stock < req.body.quantity) return next(new AppError(400, "Not enough stock"));
+
+      item.quantity = req.body.quantity;
+      await cart.save();
+      return sendResponse(200, cart, res, `Cart item updated successfully`);
+    }
+
     const doc = await Model.findByIdAndUpdate(req.params.id, req.body, {
       runValidators: true,
       new: true,
@@ -73,8 +108,21 @@ exports.updateOne = (Model, options) =>
     sendResponse(200, doc, res, `${Model.modelName} updated successfully!`);
   });
 
-exports.deleteOne = (Model) =>
+exports.deleteOne = (Model, options) =>
   catchAsync(async (req, res, next) => {
+    if (options?.cartLogic) {
+      const cart = await Model.findOne({ user: req.user.id });
+      if (!cart) return next(new AppError(404, "Cart not found"));
+
+      const item = cart.items.id(req.params.id);
+      if (!item) return next(new AppError(404, "Item not found"));
+
+      cart.items.pull(item.id);
+      await cart.save();
+
+      return sendResponse(200, cart, res, `Cart item deleted successfully`);
+    }
+
     const doc = await Model.findByIdAndDelete(req.params.id);
 
     if (!doc) return next(new AppError(404, `no ${Model.modelName} found with this ID.`));
@@ -82,8 +130,15 @@ exports.deleteOne = (Model) =>
     sendResponse(204, null, res);
   });
 
-exports.deleteAll = (Model) =>
+exports.deleteAll = (Model, options) =>
   catchAsync(async (req, res, next) => {
-    await Model.deleteMany();
-    sendResponse(204, null, res);
+    if (options?.cartLogic) {
+      const cart = await Model.find({ user: req.user.id });
+      if (!cart) return next(new AppError(404, "Cart not found"));
+      await Model.findByIdAndDelete(cart._id);
+      sendResponse(204, null, res);
+    } else {
+      await Model.deleteMany();
+      sendResponse(204, null, res);
+    }
   });
